Important update, corrected the problem of the previous commit.
There is a huge problem in the previous commit. The usage scenario of that function should be in the "event loop", not in the "slot function". Note that the "slot function" is not necessarily executed in the "main thread", it may also be executed in the "sub-thread", and what we care about is actually the execution in the "event loop of the main thread".